Removing all enums from DBCreationManager and create separet txt file for these tables and create method for parthing this file

+++ src/main/java/edu/softserveinc/healthbody/db/DBCreationManager.java
@@ -1,94 +1,23 @@
 package edu.softserveinc.healthbody.db;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.Arrays;
+import java.util.List;
 
 import org.slf4j.LoggerFactory;
 import org.slf4j.Logger;
 
 public class DBCreationManager {
-	public static enum TableQueries {
-		USERS_TABLE("CREATE TABLE IF NOT EXISTS \"users\"("
-				+ "id_user serial primary key,"
-				+ "login varchar(50),"
-				+ "password varchar(25),"
-				+ "firstname varchar(50),"
-				+ "lastname varchar(50),"
-				+ "\"e-mail\" varchar(50),"
-				+ "age bigint,"
-				+ "weight real,"
-				+ "gender \"char\","
-				+ "health varchar(50),"
-				+ "avatar varchar(50),"
-				+ "google_field varchar(150),"
-				+ "id_role bigint,"
-				+ "status varchar(50),"
-				+ "isDisabled boolean)"),
-		ROLE_TABLE("CREATE TABLE IF NOT EXISTS \"roles\"("
-				+ "id_role serial primary key,"
-				+ "name varchar(50),"
-				+ "description varchar(200))"),
-		GROUPS_TABLE("CREATE TABLE IF NOT EXISTS \"groups\"("
-				+ "id_group serial primary key,"
-				+ "name varchar(50),"
-				+ "count bigint,"
-				+ "description varchar(50),"
-				+ "scoreGroup varchar(50),"
-				+ "status varchar(50))"),
-		COMPETION_TABLE("CREATE TABLE IF NOT EXISTS \"competitions\"("
-				+ "id_competition serial primary key,"
-				+ "name varchar(50),"
-				+ "description varchar(200),"
-				+ "start date,"
-				+ "finish date,"
-				+ "id_criteria bigint)"),
-		AWARD_TABLE("CREATE TABLE IF NOT EXISTS \"awards\"("
-				+ "id_award serial primary key,"
-				+ "name varchar(50))"),
-		CRITERIA_TABLE("CREATE TABLE IF NOT EXISTS \"criteria\"("
-				+ "id_criteria serial primary key,"
-				+ "name varchar(50),"
-				+ "metrics real,"
-				+ "get_google varchar(200))"),
-		USER_GROUP_TABLE("CREATE TABLE IF NOT EXISTS \"usergroups\"("
-				+ "id_user_group serial primary key,"
-				+ "id_user bigint,"
-				+ "id_group bigint,"
-				+ "FOREIGN KEY (id_group)  REFERENCES \"groups\" (id_group),"
-				+ "FOREIGN KEY (id_user) REFERENCES \"users\" (id_user))"),
-		GROUP_COMPETITION_TABLE("CREATE TABLE IF NOT EXISTS \"groupcompetitions\"("
-				+ "id_group_competition serial primary key,"
-				+ "id_group bigint,"
-				+ "id_competition bigint,"
-				+ "FOREIGN KEY (id_group)  REFERENCES \"groups\" (id_group),"
-				+ "FOREIGN KEY (id_competition) REFERENCES \"competitions\" (id_competition))"),
-		USER_COMPETITION_TABLE("CREATE TABLE IF NOT EXISTS \"usercompetitions\"("
-				+ "id_user_competition serial primary key,"
-				+ "id_user bigint,"
-				+ "id_competition bigint,"
-				+ "id_award bigint,"
-				+ "user_score bigint,"
-				+ "time_received varchar(50),"
-				+ "FOREIGN KEY (id_user) REFERENCES \"users\" (id_user),"
-				+ "FOREIGN KEY (id_competition)  REFERENCES \"competitions\" (id_competition),"
-				+ "FOREIGN KEY (id_award) REFERENCES \"awards\" (id_award))"),
-		META_DATA_TABLE("CREATE TABLE IF NOT EXISTS \"metadata\"("
-				+ "id_metadata serial primary key,"
-				+ "last_synch varchar(50))");
-
-		private String table;
-
-		private TableQueries(String table) {
-			this.table = table;
-		}
-
-		@Override
-		public String toString() {
-			return this.table;
-		}
-	}
 
 	private static Logger logger = LoggerFactory.getLogger(DBCreationManager.class.getName());
+	private static final ClassLoader LOADER = Thread.currentThread().getContextClassLoader();
+	private static final String PATH_FILE = "tables.txt";
+	private static final String TABLES_SPLIT = ";";
 	private static volatile DBCreationManager instance = null;
 
 	private DBCreationManager() {
@@ -107,9 +36,14 @@ public class DBCreationManager {
 
 	public boolean dropDatabase(Statement statement, String databaseName) throws SQLException {
 		boolean result = false;
-		String deleteConnectionsQuery = "SELECT pg_terminate_backend(pg_stat_activity.pid) FROM pg_stat_activity WHERE pg_stat_activity.datname = \'"
-				+ databaseName + "\' AND pid <> pg_backend_pid();";
-		result = statement.execute(deleteConnectionsQuery + "DROP DATABASE " + databaseName + ";");
+		statement.execute("select datname from pg_catalog.pg_database where datname = \'" + databaseName + "\';");
+		if (statement.getResultSet().next()){
+			String deleteConnectionsQuery = "SELECT pg_terminate_backend(pg_stat_activity.pid) FROM pg_stat_activity WHERE pg_stat_activity.datname = \'"
+					+ databaseName + "\' AND pid <> pg_backend_pid();";
+			result = statement.execute(deleteConnectionsQuery + "DROP DATABASE " + databaseName + ";");			
+		} else {
+			logger.info("Cannot delete database "+databaseName+", because this datatbase is absent!");
+		}
 		return result;
 	}
 
@@ -131,6 +65,28 @@ public class DBCreationManager {
 		boolean result = false;
 		result = statement.execute(tableQuery);
 		return result;
+	}
+	
+	public List<String> getListOfQueries() {
+		List<String> queries = null;
+		String row;
+		BufferedReader bufReader = null;
+		StringBuilder sb = new StringBuilder();
+		try {
+			InputStream is = LOADER.getResourceAsStream(PATH_FILE);
+			if (is != null){
+				bufReader = new BufferedReader(new InputStreamReader(is));
+				while ((row = bufReader.readLine()) != null) {
+					sb.append(row);
+				}
+				queries = Arrays.asList(sb.toString().split(TABLES_SPLIT));				
+			} else {
+				logger.error("Problem with file access!");
+			}
+		} catch (IOException e) {
+			logger.error("Cannot access to file "+PATH_FILE, e);
+		}		
+		return queries;
 	}
